feat(text-editor): rename text-editor and text-editor-form-component
Rename limel-text-editor to limel-prosemirror-adapter
Rename text-editor-form-component to limel-text-editor
