+++ src/components/text-editor/examples/prosemirror-adapter-with-custom-menu.tsx
@@ -1,0 +1,32 @@
+import { LimelProsemirrorAdapterCustomEvent } from '@limetech/lime-elements';
+import { Component, h, State } from '@stencil/core';
+/**
+ * Example with custom menu
+ */
+@Component({
+    tag: 'limel-example-prosemirror-adapter-with-custom-menu',
+    shadow: true,
+})
+export class ProsemirrorAdapterWithCustomMenuExample {
+    @State()
+    private text: { html: string } = { html: '' };
+
+    public render() {
+        // const menuItems: EditorPluginButton[] = ['undo', 'redo'];
+
+        return [
+            <limel-prosemirror-adapter
+                onChange={this.handleChange}
+                // menuItems={menuItems}
+            />,
+            <limel-example-value value={this.text} />,
+        ];
+    }
+
+    private handleChange = (
+        event: LimelProsemirrorAdapterCustomEvent<{ html: string }>,
+    ): void => {
+        event.stopPropagation();
+        this.text = event.detail;
+    };
+}
