+++ src/components/text-editor/prosemirror-adapter.scss
@@ -1,0 +1,50 @@
+@use '../../style/internal/shared_input-select-picker.scss';
+
+:host(limel-text-editor) {
+    isolation: isolate;
+    display: block;
+}
+
+* {
+    box-sizing: border-box;
+}
+
+.ProseMirror-menubar-wrapper {
+    transition: border 0.2s ease;
+    display: grid;
+    grid-template-rows: auto 1fr;
+    border-radius: 0.25rem;
+    border: 1px solid;
+    border-color: shared_input-select-picker.$lime-text-field-outline-color;
+
+    &:hover {
+        border-color: shared_input-select-picker.$lime-text-field-outline-color--hovered;
+    }
+
+    &:focus-within {
+        border-color: shared_input-select-picker.$lime-text-field-outline-color--focused;
+    }
+}
+
+.ProseMirror-textblock-dropdown {
+    min-width: 3em;
+}
+
+.ProseMirror-tooltip {
+    .ProseMirror-menu {
+        width: -webkit-fit-content;
+        width: fit-content;
+        white-space: pre;
+    }
+}
+
+@import './partial-styles/prosemirror.scss';
+@import './partial-styles/prosemirror-menu.scss';
+
+@import '../markdown/partial-styles/blockquotes';
+@import '../markdown/partial-styles/body-text';
+@import '../markdown/partial-styles/definition-lists';
+@import '../markdown/partial-styles/headings';
+@import '../markdown/partial-styles/lists';
+@import '../markdown/partial-styles/pre-code';
+@import '../markdown/partial-styles/tables';
